Have a better profile-agnostic report introduction
The "authority table" (AT) may have been mission-specific wording for a
specific time, but it is no longer relevant. The European Union's ATs
are indeed encoded as SKOS vocabularies, so that's what we can call
them, with the assumption that our tool can diff other SKOS vocabs as
well. Same goes for RDFS/OWL (we put a slash to indicate it can be any
subset of pure RDFS and OWL) and ShACL shapes.