Fix deletion queries returning all values
When the deletions graph contains deleted properties that have multiple
values, ALL of those values are returned in the deletion queries,
instead of just the ones that were deleted.

It turns out that in the deletion BGPs the values are always unbound
because of the anonymous variable `[]`. This brings in all values when
the final BGP for returning optional labels matches on all of them,
because of this loose binding.

In order to avoid this, simply use the named variable that is already
used in the last BGP. Doing this should have no side-effects.