test(data): update list of changes in OWL diff sample
Recheck the test data files against the latest export of the same
diff/report, mostly reordering the changes and making some observations.

Redundancies not observed any more:

- class **epo:AdHocChanel** moved `rdfs:isDefinedBy` to **epo:AwardCriterion**

- objectProperty **epo:exposesChannel** moved `rdfs:isDefinedBy` to **epo:followsRuleSetBy**

- datatypeProperty **epo:describesProfessionRelevantLaw** moved `rdfs:isDefinedBy` to **epo:describesObjectiveParticipationRules**

These may be because of updates to the tool since the last time these
descriptions were prepared (as these are redundancies they are not
important), or they could be regressions in some manner, related to the
_moved_ change type (even if they are redundancies, we may expect them
to show up like other redundancies).

Observations (re)confirmed:

- change of language to null (deleted lang) is not captured (as an update)

- change of datatype property to object property is not captured at all (but should be captured as one deleted and the other added at least)

- change of language is not captured (as an update)

- the _move_ change type is not represented in this diff sample